+++ DependencyInjection/c975LPurchaseCreditsExtension.php
@@ -15,11 +15,13 @@ use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
+/**
+ * DI Extension Class
+ * @author Laurent Marquet <laurent.marquet@laposte.net>
+ * @copyright 2018 975L <contact@975l.com>
+ */
 class c975LPurchaseCreditsExtension extends Extension
 {
-    /**
-     * {@inheritdoc}
-     */
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new YamlFileLoader(
@@ -33,6 +35,7 @@ class c975LPurchaseCreditsExtension extends Extension
 
         $container->setParameter('c975_l_purchase_credits.creditsNumber', $processedConfig['creditsNumber']);
         $container->setParameter('c975_l_purchase_credits.creditsPrice', $processedConfig['creditsPrice']);
+        $container->setParameter('c975_l_purchase_credits.gdpr', $processedConfig['gdpr']);
         $container->setParameter('c975_l_purchase_credits.currency', strtoupper($processedConfig['currency']));
         $container->setParameter('c975_l_purchase_credits.vat', $processedConfig['vat'] * 100);
         $container->setParameter('c975_l_purchase_credits.live', $processedConfig['live']);