+++ Security/TransactionVoter.php
@@ -14,14 +14,40 @@ use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\Voter;
 use c975L\PurchaseCreditsBundle\Entity\Transaction;
 
+/**
+ * Voter for Transaction access
+ * @author Laurent Marquet <laurent.marquet@laposte.net>
+ * @copyright 2018 975L <contact@975l.com>
+ */
 class TransactionVoter extends Voter
 {
+    /**
+     * @var AccessDecisionManagerInterface
+     */
     private $decisionManager;
+
+    /**
+     * The role needed to be allowed access (defined in config)
+     * @var string
+     */
     private $roleNeeded;
 
+    /**
+     * Used for access to all
+     * @var string
+     */
     public const ALL = 'all';
+
+    /**
+     * Used for access to display
+     * @var string
+     */
     public const DISPLAY = 'display';
 
+    /**
+     * Contains all the available attributes to check with in supports()
+     * @var array
+     */
     private const ATTRIBUTES = array(
         self::ALL,
         self::DISPLAY,
@@ -33,6 +59,10 @@ class TransactionVoter extends Voter
         $this->roleNeeded = $roleNeeded;
     }
 
+    /**
+     * Checks if attribute and subject are supported
+     * @return bool
+     */
     protected function supports($attribute, $subject)
     {
         if (null !== $subject) {
@@ -42,6 +72,11 @@ class TransactionVoter extends Voter
         return in_array($attribute, self::ATTRIBUTES);
     }
 
+    /**
+     * Votes if access is granted
+     * @return bool
+     * @throws \LogicException
+     */
     protected function voteOnAttribute($attribute, $subject, TokenInterface $token)
     {
         //Defines access rights
@@ -57,7 +92,10 @@ class TransactionVoter extends Voter
         throw new \LogicException('Invalid attribute: ' . $attribute);
     }
 
-    //Checks if user is owner or has admin rights
+    /**
+     * Checks if user is owner or has admin rights
+     * @return bool
+     */
     private function isOwner($token, Transaction $transaction)
     {
         if ($this->decisionManager->decide($token, array('ROLE_USER'))) {
@@ -67,7 +105,10 @@ class TransactionVoter extends Voter
         return false;
     }
 
-    //Checks if user has admin rights
+    /**
+     * Checks if user has admin rights
+     * @return bool
+     */
     private function isAdmin($token)
     {
         return $this->decisionManager->decide($token, array($this->roleNeeded));