+++ Service/PurchaseCreditsService.php
@@ -9,74 +9,101 @@
 
 namespace c975L\PurchaseCreditsBundle\Service;
 
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
+use c975L\PaymentBundle\Entity\Payment;
+use c975L\ServicesBundle\Service\ServiceToolsInterface;
+use c975L\PurchaseCreditsBundle\Entity\PurchaseCredits;
+use c975L\PurchaseCreditsBundle\Service\TransactionServiceInterface;
+use c975L\PurchaseCreditsBundle\Service\PurchaseCreditsServiceInterface;
+use c975L\PurchaseCreditsBundle\Service\Email\PurchaseCreditsEmailInterface;
 
-class PurchaseCreditsService
+/**
+ * PurchaseCreditsService class
+ * @author Laurent Marquet <laurent.marquet@laposte.net>
+ * @copyright 2017 975L <contact@975l.com>
+ */
+class PurchaseCreditsService implements PurchaseCreditsServiceInterface
 {
+    /**
+     * Stores ContainerInterface
+     * @var ContainerInterface
+     */
     private $container;
+
+    /**
+     * Stores EntityManagerInterface
+     * @var EntityManagerInterface
+     */
     private $em;
-    private $emailService;
-    private $paymentService;
+
+    /**
+     * Stores PurchaseCreditsEmailInterface
+     * @var PurchaseCreditsEmailInterface
+     */
+    private $purchaseCreditsEmail;
+
+    /**
+     * Stores ServiceToolsInterface
+     * @var ServiceToolsInterface
+     */
+    private $serviceTools;
+
+    /**
+     * Stores current Request
+     * @var RequestStack
+     */
     private $request;
-    private $router;
-    private $templating;
+
+    /**
+     * Stores TransactionServiceInterface
+     * @var TransactionServiceInterface
+     */
     private $transactionService;
 
+    /**
+     * Stores TranslatorInterface
+     * @var TranslatorInterface
+     */
+    private $translator;
+
     public function __construct(
-        \Symfony\Component\DependencyInjection\ContainerInterface $container,
-        \Doctrine\ORM\EntityManager $em,
-        \c975L\EmailBundle\Service\EmailService $emailService,
-        \c975L\PaymentBundle\Service\PaymentService $paymentService,
-        \Symfony\Component\HttpFoundation\RequestStack $requestStack,
-        \Symfony\Component\Routing\Generator\UrlGeneratorInterface $router,
-        \Twig_Environment $templating,
-        \c975L\PurchaseCreditsBundle\Service\TransactionService $transactionService
-    ) {
+        ContainerInterface $container,
+        EntityManagerInterface $em,
+        PurchaseCreditsEmailInterface $purchaseCreditsEmail,
+        ServiceToolsInterface $serviceTools,
+        RequestStack $requestStack,
+        TransactionServiceInterface $transactionService,
+        TranslatorInterface $translator
+    )
+    {
         $this->container = $container;
         $this->em = $em;
-        $this->emailService = $emailService;
-        $this->paymentService = $paymentService;
+        $this->purchaseCreditsEmail = $purchaseCreditsEmail;
+        $this->serviceTools = $serviceTools;
         $this->request = $requestStack->getCurrentRequest();
-        $this->router = $router;
-        $this->templating = $templating;
         $this->transactionService = $transactionService;
+        $this->translator = $translator;
     }
 
-    //Adds credits
-    public function add($payment)
+    /**
+     * {@inheritdoc}
+     */
+    public function create()
     {
-        $action = (array) json_decode($payment->getAction());
-        if (array_key_exists('addCredits', $action)) {
-            //Gets the user
-            $user = $this->em
-                ->getRepository($this->container->getParameter('c975_l_purchase_credits.userEntity'))
-                ->findOneById($payment->getUserId());
+        $purchaseCredits = new PurchaseCredits();
+        $purchaseCredits
+            ->setUserIp($this->request->getClientIp())
+        ;
 
-            //Adds Transaction
-            $this->transactionService->add($payment, $action['addCredits'], $user);
-
-            //Set payment as finished
-            $payment->setFinished(true);
-            $this->em->persist($payment);
-            $this->em->flush();
-
-            //Sends email
-            $this->sendEmail($payment, $action['addCredits'], $user);
-
-            //Creates flash
-            $flash = $this->container->get('translator')->trans('text.credits_purchased', array('%credits%' => $action['addCredits']), 'purchaseCredits');
-            $this->request->getSession()
-                ->getFlashBag()
-                ->add('success', $flash)
-                ;
-
-            return true;
-        }
-
-        return false;
+        return $purchaseCredits;
     }
 
-    //Gets prices for credits
+    /**
+     * {@inheritdoc}
+     */
     public function getPrices()
     {
         $prices = array();
@@ -87,11 +114,12 @@ class PurchaseCreditsService
         return $prices;
     }
 
-    //Gets prices choices labels for credits
-    public function getPricesChoices($prices)
+    /**
+     * {@inheritdoc}
+     */
+    public function getPricesChoice()
     {
-        //Defines choices labels
-        $translator = $this->container->get('translator');
+        $prices = $this->getPrices();
         $pricesChoices = array();
         $creditReferencePrice = key($prices) / reset($prices);
         $format = new \NumberFormatter('en_EN' . '@currency=' . $this->container->getParameter('c975_l_purchase_credits.currency'), \NumberFormatter::CURRENCY);
@@ -100,7 +128,7 @@ class PurchaseCreditsService
         foreach ($prices as $key => $value) {
             //Calculates the discount (if one) based on the ratio of the first $price entry
             $discount = (1 - ($value / ($key / $creditReferencePrice))) * 100;
-            $label = $key . ' ' . $translator->transChoice('label.credits', $key, array(), 'purchaseCredits') . ' - ' . $value . ' ' . $currencySymbol;
+            $label = $key . ' ' . $this->translator->transChoice('label.credits', $key, array(), 'purchaseCredits') . ' - ' . $value . ' ' . $currencySymbol;
             $label = $discount != 0 ? $label . ' (-' . $discount . ' %)' : $label;
             $pricesChoices[$label] = $key;
         }
@@ -108,109 +136,47 @@ class PurchaseCreditsService
         return $pricesChoices;
     }
 
-    //Gets the Terms of sales url
-    public function getTosUrl()
+    /**
+     * {@inheritdoc}
+     */
+    public function define(PurchaseCredits $purchaseCredits)
     {
-        return $this->getUrl($this->container->getParameter('c975_l_purchase_credits.tosUrl'));
+        $purchaseCredits
+            ->setAmount($this->getPrices()[$purchaseCredits->getCredits()] * 100)
+            ->setCurrency($this->container->getParameter('c975_l_purchase_credits.currency'))
+        ;
     }
 
-    //Gets the Terms of sales PDF
-    public function getTosPdf()
+    /**
+     * {@inheritdoc}
+     */
+    public function validate(Payment $payment)
     {
-        $tosPdfUrl = $this->getUrl($this->container->getParameter('c975_l_purchase_credits.tosPdf'));
+        $action = (array) json_decode($payment->getAction());
 
-        //Gets the content of TermsOfSales PDF
-        if ($tosPdfUrl !== null) {
-            $tosPdf = file_get_contents($tosPdfUrl);
-            $filenameTos = $this->container->get('translator')->trans('label.terms_of_sales_filename', array(), 'purchaseCredits') . '.pdf';
-            return array($tosPdf, $filenameTos, 'application/pdf');
+        if (array_key_exists('addCredits', $action)) {
+            //Gets the user
+            $user = $this->em
+                ->getRepository($this->container->getParameter('c975_l_purchase_credits.userEntity'))
+                ->findOneById($payment->getUserId());
+
+            //Adds Transaction + user's credits
+            $this->transactionService->add($payment, $action['addCredits'], $user);
+
+            //Set payment as finished
+            $payment->setFinished(true);
+            $this->em->persist($payment);
+            $this->em->flush();
+
+            //Sends email
+            $this->purchaseCreditsEmail->send($payment, $action['addCredits'], $user);
+
+            //Creates flash
+            $this->serviceTools->createFlash('purchasedCredits', 'text.credits_purchased', 'success', array('%credits%' => $action['addCredits']));
+
+            return true;
         }
 
-        return null;
-    }
-
-    //Defines the url
-    public function getUrl($data)
-    {
-        //Calculates the url if a Route is provided
-        if (false !== strpos($data, ',')) {
-            $routeData = $this->getUrlFromRoute($data);
-            $url = $this->router->generate($routeData['route'], $routeData['params'], UrlGeneratorInterface::ABSOLUTE_URL);
-        //An url has been provided
-        } elseif (false !== strpos($data, 'http')) {
-            $url = $data;
-        //Not valid data
-        } else {
-            $url = null;
-        }
-
-        return $url;
-    }
-
-    //Gets url from a Route
-    public function getUrlFromRoute($route)
-    {
-        //Gets Route
-        $routeValue = trim(substr($route, 0, strpos($route, ',')), "\'\"");
-
-        //Gets parameters
-        $params = trim(substr($route, strpos($route, '{')), "{}");
-        $params = str_replace(array('"', "'"), '', $params);
-        $params = explode(',', $params);
-
-        //Caculates url
-        $paramsArray = array();
-        foreach($params as $value) {
-            $parameter = explode(':', $value);
-            $paramsArray[trim($parameter[0])] = trim($parameter[1]);
-        }
-
-        return array(
-            'route' => $routeValue,
-            'params' => $paramsArray
-        );
-    }
-
-    //Defines payment for Credits purchased
-    public function payment($purchaseCredits, $userId)
-    {
-        $paymentData = array(
-            'amount' => $purchaseCredits->getAmount(),
-            'currency' => $purchaseCredits->getCurrency(),
-            'action' => json_encode(array('addCredits' => $purchaseCredits->getCredits())),
-            'description' => $this->container->get('translator')->trans('label.purchase_credits', array(), 'purchaseCredits') . ' (' . $purchaseCredits->getCredits() . ')',
-            'userId' => $userId,
-            'userIp' => $this->request->getClientIp(),
-            'live' => $this->container->getParameter('c975_l_purchase_credits.live'),
-            'returnRoute' => 'purchasecredits_payment_done',
-            'vat' => $this->container->getParameter('c975_l_purchase_credits.vat'),
-            );
-        $this->paymentService->create($paymentData);
-    }
-
-    //Sends email for GiftVoucher purchased
-    public function sendEmail($payment, $credits, $user)
-    {
-        //Gets the PDF of Terms of sales
-        $tosPdf = $this->getTosPdf();
-
-        //Sends email
-        $body = $this->templating->render('@c975LPurchaseCredits/emails/purchase.html.twig', array(
-            'payment' => $payment,
-            'credits' => $credits,
-            'userCredits' => $user->getCredits(),
-            'live' => $this->container->getParameter('c975_l_purchase_credits.live'),
-             '_locale' => $this->request->getLocale(),
-            ));
-        $emailData = array(
-            'subject' => $this->container->get('translator')->trans('label.purchased_credits', array('%credits%' => $credits), 'purchaseCredits'),
-            'sentFrom' => $this->container->getParameter('c975_l_email.sentFrom'),
-            'sentTo' => $user->getEmail(),
-            'replyTo' => $this->container->getParameter('c975_l_email.sentFrom'),
-            'body' => $body,
-            'attach' => array($tosPdf),
-            'ip' => $this->request->getClientIp(),
-            );
-        $this->emailService->send($emailData, true);
+        return false;
     }
 }